+++ docs/version_history.md
@@ -9,6 +9,7 @@
 - additional LRO LAMP RDR table/image pointers
 - raw images of 9P/TEMPEL 1 from the Keck Observatory
 - Deep Space 1 safed MICAS and PEPE datasets
+- Lunar Prospector ancillary ascii tables
 
 ### Changed